v2v: use datavolume instead of PVC
Hot fix only.
The PVC should be enough but recent validation within UI expects the PVC to be already
created if used in the StorageTab of the Create VM Wizard.

The v2v flow will create the PVC during final step of the wizard, right before
createing the Conversion POD.