v2v-conversion-pod: the temp disk is always of volumeMode=Filesystem
To make flows in the conversion pod easier, the temporary disk is requested with
the Filesystem volumeMode no matter of the kubevirt-storage-class-defaults values.

The temporary space for the conversion pod is requested on the default storage class.
The "Filesystem" volumeMode seems to be safe, supported by all storage classes.

The target disks for the v2v stay untouched, this change affects just the temporary space.

Signed-off-by: Marek Libra <marek.libra@gmail.com>